% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetREWS.R
\name{AreaBelow}
\alias{AreaBelow}
\title{Calculate the area below a certain level in a rotor disk}
\usage{
AreaBelow(h, R)
}
\arguments{
\item{h}{the height with respect to the bottom of the disk}

\item{R}{the disk radius}
}
\value{
a single value of the area in the circle, below height h.
}
\description{
\code{AreaBelow} is a utility function to calculate the  the area below a
certain height in the rotor disk. The height is given with respect to the
bottom of the rotor disk.
}
\seealso{
\code{\link{SliceArea}}
}

