% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanPlot.R
\name{cleanPlot}
\alias{cleanPlot}
\title{Generate a clean plot that is more-or-less publication ready}
\usage{
cleanPlot(p, base.size = 10, code.dir = dirname(sys.frame(1)$ofile),
  axis.colour = "grey50", grid.colour = "grey90",
  axis.text.colour = "grey33", plot.background = "white",
  panel.background = "white")
}
\arguments{
\item{p}{the ggplot plot to be processed}

\item{base.size}{the default text size in points}
}
\value{
p the cleaned plot
}
\description{
\code{cleanPlot} tidies up standard ggplot2 graphics.
}
\seealso{
\code{\link{ggplot2}}
}

