% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetErrorMetrics.R
\name{GetErrorMetrics}
\alias{GetErrorMetrics}
\title{Calculate a range of error metrics for a data set}
\usage{
GetErrorMetrics(y, ymod)
}
\arguments{
\item{y}{the actual (observed) value}

\item{ymod}{the modelled (predicted) value}
}
\value{
metrics a data.frame with columns for the metric and value
}
\description{
\code{GetErrorMetrics} calculates a range of error metrics.
}
\seealso{
\code{\link{PearsonCorrelation}, \link{RMSE}, \link{MAE},
  \link{MaxAE}, \link{SAE}}
}

