% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetREWS.R
\name{GetREWS}
\alias{GetREWS}
\title{Calculate rotor-equivalent wind speed from a wind speed profile}
\usage{
GetREWS(u, z, zhub, dia)
}
\arguments{
\item{u}{a vector of wind speeds}

\item{z}{a vector of heights}

\item{zhub}{the turbine hub height}

\item{dia}{the turbine rotor diameter}
}
\value{
a single value of the rotor equvalent wind speed
}
\description{
\code{GetREWS} calculates the rotor-equivalent wind speed using the method set out in "Accounting for the speed shear in wind turbine power performance measurement" (Wind Energy, 2011).
}
\seealso{
\code{\link{GetErrorMetrics}}
}

