% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetTOR.R
\name{GetTOR}
\alias{GetTOR}
\title{Get the operating region of a wind turbine}
\usage{
GetTOR(ws, ws.cutin = 0, ws.rated, ws.cutout = Inf)
}
\arguments{
\item{ws}{a wind speed}

\item{ws.cutin}{the speed above which the turbine is operating}

\item{ws.rated}{the speed above which the turbine runs at rated power}

\item{ws.cutout}{the speed above which the turbine is shut down for safety
(or other) reasons}
}
\value{
TOR the turbine operating region (factor = c("I", "II", or "III"))
}
\description{
\code{GetTOR} calculates the operating range of a wind
turbine based on the wind speed and the turbine cutin, rated and cutout wind
speeds.
}

