% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MAE.R
\name{MAE}
\alias{MAE}
\title{Calculate the mean absolute error for a data set}
\usage{
MAE(y, ymod)
}
\arguments{
\item{y}{the actual (observed) value}

\item{ymod}{the modelled (predicted) value}
}
\value{
a single value of the MAE
}
\description{
\code{MAE} calculates the mean of the absolute difference between two values
(predicted and actual), or the MAE. Any observations or predicted values that
are NA are removed.
}
\seealso{
\code{\link{GetErrorMetrics}}
}

