% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NRELPlot.R
\name{NRELPlot}
\alias{NRELPlot}
\title{Generate a clean plot that is more-or-less publication ready}
\usage{
NRELPlot(p, base.size = 10, code.dir = dirname(sys.frame(1)$ofile),
  axis.colour = "grey50", grid.colour = "grey75",
  axis.text.colour = "black", axis.title.colour = "black",
  plot.background = "white", panel.background = "grey95")
}
\arguments{
\item{p}{the ggplot plot to be processed}

\item{base.size}{the default text size in points}
}
\value{
p the cleaned plot
}
\description{
\code{cleanPlot} tidies up standard ggplot2 graphics.
}
\seealso{
\code{\link{ggplot2}}
}

