% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCfromParam.R
\name{PCfromParam}
\alias{PCfromParam}
\title{Generate a power curve based on parameters}
\usage{
PCfromParam(param, ws = seq(3, 30, 0.5))
}
\arguments{
\item{param}{is a data.frame that contains paramaeters that describe a wind turbine's performance}
}
\description{
Generate a power curve based on parameters
}

