% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCQueryTurbineModel.R
\name{PCQueryTurbineModel}
\alias{PCQueryTurbineModel}
\title{Get an estimate of turbine power production from a wind speed-power curve}
\usage{
PCQueryTurbineModel(power.curve, ws = NULL, power = NULL)
}
\arguments{
\item{power.curve}{a power curve object containing information about wind
speeds and power}

\item{ws}{the wind speed (can be a vector)}

\item{power}{the power (can be a vector)}
}
\value{
if ws is input, a vector of power at wind speed ws

if power is input, a vector of the lowest wind speed to achieve that
 power
}
\description{
\code{PCQueryTurbineModel} queries the power curve to find the power expected at a
given wind speed, if the wind speed is input. If the power is input, the wind
speed will be estimated from the power. \code{PCQueryTurbineModel} uses
\code{approx} to perform a linear interpolation between the input power curve
data points to the point of interest. Wind speeds that exceed the range of
data in the input power curve are assumed to have 0 mean power.
}
\note{
either ws or power should be given as an input.
}
\seealso{
Other Power curve methods: \code{\link{PCTrainTest}};
  \code{\link{PCTrainTurbineModel}};
  \code{\link{PCwsAdjDens}}; \code{\link{plotPC}}
}

