% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCTrainTestTurbineModel.R
\name{PCTrainTest}
\alias{PCTrainTest}
\title{Train and test a power curve model}
\usage{
PCTrainTest(train.predictors, train.response, test.predictors, test.response)
}
\arguments{
\item{train.predictors}{a data.frame containing the predictors that will be
used to train the model}

\item{train.response}{a vector containing the response of the turbine (this
could be power, loads, ...)}

\item{test.predictors}{a data.frame containing a new set of predictors that
will be use to create a new response matrix}

\item{test.response}{a vector of observed responses}
}
\value{
a list containing \describe{\item{model}{the power curve}
  \item{metrics}{a data.frame containing different error and performance
  metrics}}
}
\description{
\code{PCTrainTest} trains and then tests a power curve model of a
wind turbine using actual observations. The function returns basic error and
performance (CPU) metrics. This function is intended for assessing the
accuracy of a particular model.
}
\seealso{
Other Power curve methods: \code{\link{PCQueryTurbineModel}};
  \code{\link{PCTrainTurbineModel}};
  \code{\link{PCwsAdjDens}}; \code{\link{plotPC}}
}

