% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCTrainTurbineModel.R
\name{PCTrainTurbineModel}
\alias{PCTrainTurbineModel}
\title{Create an IEC-conformal power curve}
\usage{
PCTrainTurbineModel(ws, power, ti = rep(NA, length(ws)), power.rated = NA,
  ws.delta = 0.5, ws.cutin = 3, ws.min = 0.25,
  ws.max = (round(max(ws)/ws.delta) * ws.delta) - ws.delta)
}
\arguments{
\item{ws}{wind speed observations}

\item{power}{power observations}

\item{ti}{(optional) the observed turbulence intensity}

\item{power.rated}{(optional) the rated power of the turbine generator}

\item{ws.delta}{(optional, default = 0.5 m/s) the width of the bins}

\item{ws.cutin}{(optional, default = 3 m/s)}

\item{ws.min}{the minimum wind speed, below which power = 0}

\item{ws.max}{the maximum wind speed to bin data to}
}
\value{
a data.frame containing \describe{\item{ws.binmin}{The minimum wind
  speed in that bin} \item{ws.binmax}{the maximum wind speed in that bin}
  \item{ws.binnpoints}{The number of observations (points) in that bin}
  \item{ws.binmean}{The mean wind speed in that bin} \item{ti.binmean}{The
  mean turbulence intensity in that bin (NA if no Ti given)}
  \item{power.binmean}{The mean power in that bin}
  \item{power.binsd}{The standard deviation of the power in that bin}}
}
\description{
\code{PCTrainTurbineModel} creates a wind turbine power curve using the method
described in IEC 61400-12-1 (2005).
}
\details{
The input windspeed is assumed to have been corrected for density.
}
\seealso{
\code{\link{GetErrorMetrics}, \link{RFTrainTurbineModel}}

Other Power curve methods: \code{\link{PCQueryTurbineModel}};
  \code{\link{PCTrainTest}}; \code{\link{PCwsAdjDens}};
  \code{\link{plotPC}}
}

