% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCwsAdjDens.R
\name{PCwsAdjDens}
\alias{PCwsAdjDens}
\title{Apply a density correction to wind speed according to IEC 61400-12-1 (2005)}
\usage{
PCwsAdjDens(ws, rho, rho.ref = 1.225, rho.round = 0.05)
}
\arguments{
\item{ws}{observed wind speed}

\item{rho}{observed density}

\item{rho.ref}{the reference density to use (defaults to 1.225 kg m\eqn{^{-3}})}

\item{rho.round}{the value to round to (defaults to 0.05 kg m\eqn{^{-3}})}
}
\value{
ws.adj the adjusted wind speed
}
\description{
\code{PCwsAdjDens} applies the wind speed corection described in IEC
61400-12-1 (2005) for active power controlled turbines (i.e. pitch controlled
turbines).
}
\details{
Density is rounded to the nearest 0.05 kg m\eqn{^{-3}}, except in the region of
\code{rho.ref} \eqn{\pm} 0.05 kg m\eqn{^{-3}}.
}
\seealso{
Other Power curve methods: \code{\link{PCQueryTurbineModel}};
  \code{\link{PCTrainTest}};
  \code{\link{PCTrainTurbineModel}}; \code{\link{plotPC}}
}

