% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PearsonCorrelation.R
\name{PearsonCorrelation}
\alias{PearsonCorrelation}
\title{Calculate the Pearson's Correlation Coefficient for a dataset}
\usage{
PearsonCorrelation(y, ymod)
}
\arguments{
\item{y}{the actual (observed) value}

\item{ymod}{the modelled (predicted) value}
}
\value{
the Pearson Correlation Coefficient,. Use $statistic to access the value from the test.
}
\description{
\code{PearsonCorrelation} calculates the Pearson's correlation coeffcient between two variables. Any observations or predicted values that
are NA are removed. This can be considered the sum of the underprediction
}
\seealso{
\code{\link{GetErrorMetrics}}
}

