% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RFCheckVarImp.R
\name{RFCheckVarImp}
\alias{RFCheckVarImp}
\title{Check the importance of the variables used in the Random Forest model}
\usage{
RFCheckVarImp(model.RF)
}
\arguments{
\item{model.RF}{the trained random forest model}
}
\description{
\code{RFCheckVarImp} shows the importance of the different variables that
were input to the random forest model.
}
\seealso{
Other Random forest methods: \code{\link{RFChecknTrees}};
  \code{\link{RFQueryTurbineModel}};
  \code{\link{RFTrainTestTurbineModel}};
  \code{\link{RFTrainTurbineModel}}
}

