% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RFQueryTurbineModel.R
\name{RFQueryTurbineModel}
\alias{RFQueryTurbineModel}
\title{Use a random forest model to predict response to forcing conditions}
\usage{
RFQueryTurbineModel(predictors, model)
}
\arguments{
\item{predictors}{a data.frame of predictors corresponding to the data.frame
used to train the model}

\item{model}{a model object that was trained using predictors with the same names}
}
\value{
mean and standard deviation of the values from the model
}
\description{
\code{RFQueryTurbineModel} uses a trained random forest model to estimate the
response to a set of forcing conditions (predictors).
}
\seealso{
Other Random forest methods: \code{\link{RFCheckVarImp}};
  \code{\link{RFChecknTrees}};
  \code{\link{RFTrainTestTurbineModel}};
  \code{\link{RFTrainTurbineModel}}
}

