% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RFTrainTestTurbineModel.R
\name{RFTrainTestTurbineModel}
\alias{RFTrainTestTurbineModel}
\title{Train and test a random forest power model}
\usage{
RFTrainTestTurbineModel(train.predictors, train.response, test.predictors,
  test.response)
}
\arguments{
\item{train.predictors}{a data.frame containing the predictors that will be
used to train the model}

\item{train.response}{a vector containing the response of the turbine (this
could be power, loads, ...)}

\item{test.predictors}{a data.frame containing a new set of predictors that
will be use to create a new response matrix}

\item{test.response}{a vector of observed responses}
}
\value{
a list containing \describe{\item{model}{the random forest model}
  \item{metrics}{a data.frame containing different error and performance
  metrics}}
}
\description{
\code{RFTrainTestTurbineModel} trains and then tests a random forest power
model of a wind turbine using actual observations. The function returns basic
error and CPU usage metrics. This function is intended for assessing
the accuracy of a particular model.
}
\seealso{
\code{\link{GetErrorMetrics}}

Other Random forest methods: \code{\link{RFCheckVarImp}};
  \code{\link{RFChecknTrees}};
  \code{\link{RFQueryTurbineModel}};
  \code{\link{RFTrainTurbineModel}}
}

