% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RFTrainTurbineModel.R
\name{RFTrainTurbineModel}
\alias{RFTrainTurbineModel}
\title{Train a random forest model}
\usage{
RFTrainTurbineModel(predictors, response, ntree = 1001, ...)
}
\arguments{
\item{predictors}{a data.frame of predictive (forcing) values}

\item{response}{a vector of response data}

\item{ntree}{the number of trees to try out (default = 100)}
}
\value{
model.RF the resulting model
}
\description{
\code{RFTrainTurbineModel} trains a random forest model. Althrough the
package assumes that the model will be power versus some forcing conditions,
the function can be use to train a model with any combination of predictor
and response data.
}
\seealso{
Other Random forest methods: \code{\link{RFCheckVarImp}};
  \code{\link{RFChecknTrees}};
  \code{\link{RFQueryTurbineModel}};
  \code{\link{RFTrainTestTurbineModel}}
}

