% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Calculate the root mean square error for a data set}
\usage{
RMSE(y, ymod)
}
\arguments{
\item{y}{the actual (observed) value}

\item{ymod}{the modelled (predicted) value}
}
\value{
a single value of the RMSE
}
\description{
\code{RMSE} calculates the root of the mean of the square of the difference
between two values (predicted and actual), or the RMSE. Any values that are
NA are removed.
}
\seealso{
\code{\link{GetErrorMetrics}}
}

