% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAE.R
\name{SAE}
\alias{SAE}
\title{Calculate the sum of the absolute error for a data set}
\usage{
SAE(y, ymod)
}
\arguments{
\item{y}{the actual (observed) value}

\item{ymod}{the modelled (predicted) value}
}
\value{
a single value of the SAE
}
\description{
\code{SAE} calculates the sum of the absolute difference between two values
(predicted and actual), or the SAE. Any observations or predicted values that
are NA are removed.
}
\seealso{
\code{\link{GetErrorMetrics}}
}

