% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimPC.R
\name{SimPC}
\alias{SimPC}
\title{Simulate the power produced in a bin}
\usage{
SimPC(PC.param, ws, Ti, rho = 1.225, rel.tol = 0.001,
  method = "simpletrap")
}
\arguments{
\item{PC.param}{an optional data.frame containing $power.rated, which is the rated power}

\item{ws}{wind speed}

\item{Ti}{turbulence intensity}

\item{rho}{air density (required to estimate Cp)}

\item{rel.tol}{relative tolerance of the integrand, expressed as ((computed - exact) / exact)}

\item{PC.values}{a data.frame containing binned wind speed and power data}
}
\value{
sim a data.frame containing simulated wind speed and power for this
  turbine
}
\description{
\code{SimPC} is a utility function required to calculate the zero-turbulence
power curve.
}
\seealso{
Other zero-turbulence power curve methods: \code{\link{ZTQueryTurbineModel}};
  \code{\link{ZTTrainBaselinePCModel}};
  \code{\link{ZTTrainFinalTurbineModel}};
  \code{\link{ZTTrainInitTurbineModel}};
  \code{\link{ZTTrainTheoTurbineModel}};
  \code{\link{ZTTrainTiSpecificTurbineModel}}
}

