% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetREWS.R
\name{SliceArea}
\alias{SliceArea}
\title{Calculate the area of a slice through a rotor disk}
\usage{
SliceArea(z, zhub, dia)
}
\arguments{
\item{z}{the height above ground}

\item{zhub}{the height of the turbine hub}

\item{dia}{the turbine diameter}
}
\value{
a data frame containing
\describe{\item{z}{the height of the anemometer}
\item{slice.lower}{the lower extent of the slice}
\item{slice.upper}{the upper extent of the slice}
\item{Ai}{the area of this slice}
\item{Atotal}{the total area of the rotor disk, calculated from the sum of the slice areas}
\item{Ageom}{the area of the rotor disk, calculated using simple geometry}
\item{Apercent}{the area of this slice, relative to the area of the rotor disk}
}
}
\description{
\code{SliceArea} is a utility function to calculate the area of different slices through a turbine rotor disk
}
\seealso{
\code{\link{SliceArea}}
}

