% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SPE.R
\name{SPE}
\alias{SPE}
\title{Calculate the sum of the positive error for a data set}
\usage{
SPE(y, ymod)
}
\arguments{
\item{y}{the actual (observed) value}

\item{ymod}{the modelled (predicted) value}
}
\value{
a single value of the SPE
}
\description{
\code{SPE} calculates the sum of the positive difference between two values
(actual - predicted), or the SPE. Any observations or predicted values that
are NA are removed. This can be considered the sum of the underprediction
}
\seealso{
\code{\link{GetErrorMetrics}}
}

