% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TrpIntegrt.R
\name{TrpIntegrt}
\alias{TrpIntegrt}
\title{Trapezoidal integration}
\usage{
TrpIntegrt(x, y)
}
\arguments{
\item{x}{a vector of x data}

\item{y}{a vector of y data}
}
\value{
the integral under the curve described by x and y.
}
\description{
\code{TrpIntegrt} is a utility function to perform trapezoidal integration.
}

