% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ZTQueryTurbineModel.R
\name{ZTQueryTurbineModel}
\alias{ZTQueryTurbineModel}
\title{Query the zero-turbulence power curve}
\usage{
ZTQueryTurbineModel(PC.param, ws, Ti, power, rho, newws, newTi)
}
\arguments{
\item{PC.param}{a data.frame of parameters describing the initial
zero-turbulence power curve, created using \code{\link{ZTTrainFinalTurbineModel}}}

\item{ws}{wind speed}

\item{Ti}{turbulence intensity}

\item{power}{bin-averaged power}

\item{rho}{density}

\item{newTi}{the target turbulence intensity}
}
\value{
result a data.frame containing binned wind speed and power values
}
\description{
\code{ZTQueryTurbineModel} calculates power at a specific wind speed and turbilence intensity from a zero-turbulence power curve.
}
\seealso{
Other zero-turbulence power curve methods: \code{\link{SimPC}};
  \code{\link{ZTTrainBaselinePCModel}};
  \code{\link{ZTTrainFinalTurbineModel}};
  \code{\link{ZTTrainInitTurbineModel}};
  \code{\link{ZTTrainTheoTurbineModel}};
  \code{\link{ZTTrainTiSpecificTurbineModel}}
}

