\name{ZTSiteTurbineModel}
\alias{ZTSiteTurbineModel}
\title{Create the site specific power curve using the zero turbulence method}
\usage{
ZTSiteTurbineModel(df.PC.zt, ui, ti, rho = 1.225, diameter = NA)
}
\arguments{
  \item{df.PC.zt}{parameters describing the final
  zero-turbulence power curve created using
  \code{\link{ZTTrainFinalTurbineModel}}}

  \item{ui}{bin-averaged wind speeds}

  \item{ti}{bin-averaged turbulence intensity}

  \item{rho}{density}

  \item{diameter}{turbine density}
}
\value{
df.PC.zt parameters describing the final zero-turbulence
power curve
}
\description{
\code{ZTSiteTurbineModel} calculates the site-specific
power curve using the zero-turbuelnce method set out in the
draft version of IEC 61400-12-1 (2013). This should be the
fourth (and final) stage in calculating a site-specific
power curve using the zero-turbulence power curve
methdology.
}
\details{
The input to this function is generated by
\code{\link{ZTTrainFinalTurbineModel}}.
}
\seealso{
Other zero-turbulence power curve methods:
\code{\link{SimPC}};
\code{\link{ZTTrainFinalTurbineModel}};
\code{\link{ZTTrainInitTurbineModel}};
\code{\link{ZTTrainTheoTurbineModel}}
}

