% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ZTTrainFinalTurbineModel.R
\name{ZTTrainFinalTurbineModel}
\alias{ZTTrainFinalTurbineModel}
\title{Create the final zero-turbulence power curve}
\usage{
ZTTrainFinalTurbineModel(PC.param, ws, Ti, power, rho = 1.225)
}
\arguments{
\item{ws}{wind speeds}

\item{Ti}{turbulence intensity}

\item{power}{observed power}

\item{rho}{density}
}
\value{
df.PC.zt the final binned zero-turbulence power curve
}
\description{
\code{ZTTrainFinalTurbineModel} calculates the final zero-turbulence power curve
using the method set out in the draft version of IEC 61400-12-1 (2013). This
should be the third stage in calculating a site-specific power curve using
the zero-turbulence power curve methdology.
}
\details{
The input to this function is generated by \code{\link{ZTTrainTheoTurbineModel}}.

The result of this function is used as the input to \code{\link{ZTSiteTurbineModel}}.
}
\seealso{
Other zero-turbulence power curve methods: \code{\link{SimPC}};
  \code{\link{ZTQueryTurbineModel}};
  \code{\link{ZTTrainBaselinePCModel}};
  \code{\link{ZTTrainInitTurbineModel}};
  \code{\link{ZTTrainTheoTurbineModel}};
  \code{\link{ZTTrainTiSpecificTurbineModel}}
}

