% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ZTTrainInitTurbineModel.R
\name{ZTTrainInitTurbineModel}
\alias{ZTTrainInitTurbineModel}
\title{Initialize the zero-turbulence power curve}
\usage{
ZTTrainInitTurbineModel(PC.values, rho = NA * NROW(PC.values),
  diameter = NA)
}
\arguments{
\item{PC.values}{a power curve data.frame, created using
\code{\link{PCTrainTurbineModel}}}

\item{rho}{(optional) air density}

\item{diameter}{the turbine diameter}
}
\value{
parameters describing the initial
  zero-turbulence power curve
}
\description{
\code{ZTTrainInitTurbineModel} calculates the initial zero-turbulence power curve using
the method set out in the draft version of IEC 61400-12-1 (2013). This should
be the first stage in calculating a site-specific power curve using the
zero-turbulence power curve methdology. The result of this function should be
used as the input to \code{\link{ZTTrainTheoTurbineModel}}.
}
\seealso{
Other zero-turbulence power curve methods: \code{\link{SimPC}};
  \code{\link{ZTQueryTurbineModel}};
  \code{\link{ZTTrainBaselinePCModel}};
  \code{\link{ZTTrainFinalTurbineModel}};
  \code{\link{ZTTrainTheoTurbineModel}};
  \code{\link{ZTTrainTiSpecificTurbineModel}}
}

