\name{ZTTrainSiteTurbineModel}
\alias{ZTTrainSiteTurbineModel}
\title{Create the site specific power curve using the zero turbulence method}
\usage{
ZTTrainSiteTurbineModel(PC.values, TestSite, TargetSite, Ti.metric = "mean",
  diameter = NA)
}
\arguments{
  \item{PC.values}{parameters describing the final
  zero-turbulence power curve created using
  \code{\link{ZTTrainFinalTurbineModel}}}

  \item{ws}{bin-averaged wind speeds observed during power
  performance testing}

  \item{Ti}{bin-averaged turbulence intensity observed
  during power performance testing}

  \item{Ti.metric}{What statistic should be used to
  estimate the turbulence intensity at the new site? e.g.
  mean, mode}

  \item{diameter}{turbine diameter}
}
\value{
site specific power curve
}
\description{
\code{ZTSiteTurbineModel} calculates the site-specific
power curve using the zero-turbuelnce method set out in the
draft version of IEC 61400-12-1 (2013). This should be the
fourth (and final) stage in calculating a site-specific
power curve using the zero-turbulence power curve
methdology.
}
\details{
The input to this function is generated by
\code{\link{ZTTrainFinalTurbineModel}}.
}
\seealso{
Other zero-turbulence power curve methods:
\code{\link{SimPC}};
\code{\link{ZTTrainFinalTurbineModel}};
\code{\link{ZTTrainInitTurbineModel}};
\code{\link{ZTTrainTheoTurbineModel}}
}

