% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ZTTrainTheoTurbineModel.R
\name{ZTTrainTheoTurbineModel}
\alias{ZTTrainTheoTurbineModel}
\title{Initialize the zero-turbulence power curve}
\usage{
ZTTrainTheoTurbineModel(PC.param, ws, Ti, power, rho = 1.225,
  limit.p.sim = TRUE)
}
\arguments{
\item{PC.param}{a data.frame of parameters describing the initial
zero-turbulence power curve, created using \code{\link{ZTTrainInitTurbineModel}}}

\item{ws}{bin-averaged wind speeds}

\item{Ti}{bin-averaged turbulence intensity}

\item{power}{bin-averaged power}

\item{rho}{density}
}
\value{
PC.values a data.frame containing binned wind speed and power values
}
\description{
\code{ZTTrainTheoTurbineModel} calculates the theoretical zero-turbulence power curve
using the method set out in the draft version of IEC 61400-12-1 (2013). This
should be the second stage in calculating a site-specific power curve using
the zero-turbulence power curve methdology.
}
\details{
The input to this function is generated by \code{\link{ZTTrainInitTurbineModel}}.

The result of this function is used as the input to \code{\link{ZTTrainFinalTurbineModel}}.
}
\seealso{
Other zero-turbulence power curve methods: \code{\link{SimPC}};
  \code{\link{ZTQueryTurbineModel}};
  \code{\link{ZTTrainBaselinePCModel}};
  \code{\link{ZTTrainFinalTurbineModel}};
  \code{\link{ZTTrainInitTurbineModel}};
  \code{\link{ZTTrainTiSpecificTurbineModel}}
}

